Extract Recorder & Player from App
